+++ sugar/lib/core_range.dart
@@ -7,6 +7,9 @@
 /// A [Range] is a convex (contiguous) portion of a domain. It may be finitely constrained, i.e. `0 <= x < 5`,
 /// or extend to infinity, i.e. `x > 3`.
 ///
+/// It is recommended to use this library only if you require non-trivial range manipulation. A record, i.e. `(min, max)`
+/// is otherwise preferred.
+///
 /// ## Types of ranges
 /// Each end of a range may be either bound or unbounded. If bounded, a minimum or maximum value is associated with a
 /// range's end. Furthermore, a bounded range's end is either _closed_ (includes the value), or _open_ (excludes the value).