+++ sugar/lib/src/core/annotations.dart
@@ -35,6 +35,9 @@ import 'package:meta/meta_meta.dart';
 @sealed class NotTested {
   /// The reason the annotated location is not tested.
   final String because;
+  /// Creates a [NotTested] denoting that the annotate element is not tested because it is deeply integrated with a 3rd
+  /// party service/platform.
+  const NotTested.thirdPartyIntegration(): because = '3rd party integration';
   ///  Creates a [NotTested].
   const NotTested({required this.because});
 }